fix(plugins) sort plugins hash DAO topologically
This fixes an intermittent issue which makes kong not start.

The error manifested itself when a plugin has these three characteristics:

* At least two custom entities
* With at least one foreign relationship between them
* The schemas in the plugin's `daos.lua` are expressed as a hash-table
instead of as an array.

In other words we have a situation like this:
```
-- my_plugin/daos.lua
{ my_entities = {
    name = "my_entities",
    fields = {
      { id = typedefs.uuid },
    },
  }
  my_other_entities = {
    name = "my_other_entities",
    fields = {
      { id = typedefs.uuid },
      { ref = { type = "foreign", reference = "my_entities", on_delete = "cascade" } },
    },
  }
}
```

The hash-like table would be iterated over with `pairs` by the
plugin_loader, so in some cases `my_other_entities` would be loaded
before `my_entities`, provoking an error while loading
`my_other_entities.ref`.

Note that this problem can be sidestepped by using an array instead of a hash on
daos.lua. So instead of:
```
-- my_plugin/daos.lua
{ my_entities = {
    name = "my_entities", ...
  },
  my_other_entities =
    name = "my_other_entities", ...
  }
}
```
We could use this:
```
-- my_plugin/daos.lua
{ {
    name = "my_entities", ...
  },
  {
    name = "my_other_entities", ...
  }
}
```

And that would avoid the issue since when the plugin loader finds an array-like
table like that, it will always load the entities in the same order as they are
introduced on the table.

The test for this fix consists on using a hash-like table on the
`foreign-entity` custom plugin. If the fix is incorrect Kong should
ramdomly fail to start when that plugin is used in
`spec/02-integration/04-admin_api/17-foreign-entity_spec.lua`.